DATA-PROJECTS/DATA-PROJECT-3/api-vector-database/requirements.txt
flask
openai
requests
langchain
langchain-postgres
python-dotenv
langchain_openai
google-cloud-storage
pycopg
